Rename root vitest config
