+++ libs/service-registry/vitest.config.ts
@@ -1,5 +1,5 @@
 import { mergeConfig, defineProject } from "vitest/config";
 
-import rootConfig from "../../vitest.config.js";
+import rootConfig from "../../vitest.root.js";
 
 export default mergeConfig(rootConfig, defineProject({}));
